requirements.txt
python-telegram-bot==20.7
requests>=2.31.0
yt-dlp>=2025.3.31
python-dotenv==1.0.0